+++ CHANGELOG.md
@@ -4,6 +4,9 @@ seaweedchan
 ihavenoface:
 - 4chan Pass link by the style selector
 
+zixaphir:
+- Make Allow False Positives option more efficient
+
 ### 1.1.8 - 2013-05-01
 seaweedchan:
 - Fix QR not clearing on submit with Posting Success Notifications disabled